Optimization of the algorithm for question-set selection
Before this change, the algorithm for shuffle and selection of the question-set had an efficiency of O(n log n), being 'n' the length of the collection of all questions.

After this change, this algorithm is optimized using a partial Fisher-Yates algorithm, which brings an efficiency of O(k) or, in the worse case, of O(n), being 'k' the number of the desired length of the question-set, and 'n' the length of the collection of all questions.