External login disabled
Before this change, the external login was enabled but not configured. Therefore it was weird to see this when running.

After this change, the external login is not enabled, using the application login. Also any reference to external login within the account pages is commented out in the source code. It was not removed completely since this code comes from a template and it will be "re-enabled" in the future.