Send user registration confirmation email
Before this change, no email was sent to the user after the registration process. The user should click a confirmation link displayed on the application in order to confirm.

After this change, a email is sent to the user containing a confirmation link. Once the user open the link, the user's account is ready to login. The settings for sending emails is stored in a separate file (appsettings.LOCAL.json), which should be filled in with the corresponding settings in order to run the application. A template of this file is provided.